recommendation-service/requirements.txt
Django>=4.0
restframework
corsheaders
requests
kafka-python
psycopg2-binary
scikit-learn
pandas
numpy

